Разбил js код на модули
