+++ js/global-variables.js
@@ -1,0 +1,23 @@
+export const IMAGE_NUMBER = 25;
+
+export const COMMENT_ARRAY = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+export const COMMENTATOR_NAME_ARRAY = [
+  'Дима',
+  'Антон',
+  'Арина',
+  'Александр',
+  'Дмитрий',
+  'Анастасия',
+  'Игорь',
+  'Станислав',
+  'Юрий',
+  'Ярослав',
+];
